Updated packages, added new endpoints, and modified requests
Updated the `coverlet.collector` package version from `6.0.1` to `6.0.2` in `Mwh.Sample.Domain.Tests.csproj` and `Mwh.Sample.Repository.Tests.csproj` files. Added a new POST endpoint `/employees` in `Program.cs` that accepts an `EmployeeDto` object and uses the `IEmployeeService` to save the employee. Also added a new static class `EmployeeGroupEndpoints` in `Program.cs` that maps GET and DELETE operations for the `IEmployeeService`. Replaced the `GET` request for `/weatherforecast` with a `GET` request for `/employees/` in the `SampleMinimalApi.http` file.